save intermediate results and change value type to Float in assetOwnersWithAssetCounts field in AssetRelationship

+++ src/main/java/fund/data/assets/dto/asset/exchange/FirstBuyFixedRateBondDTO.java
@@ -41,7 +41,7 @@ public class FirstBuyFixedRateBondDTO {
     private Integer assetCount;
 
     @NotNull
-    private Map<String, Double> assetOwnersWithAssetCounts;
+    private Map<String, Float> assetOwnersWithAssetCounts;
 
     @NotNull
     private Long accountID;
